Run tests in a database transaction
(Arguably this would tell a better story with multiple commits that build on each other.)

First, add a UserService and a spec demonstrating its use.

The test will fail if run twice, because the first run will leave a record in the test database that the second run will conflict with.

Second, add the infrastructure to run jest tests in a database transaction.